requirements.txt
paho-mqtt
quart
mysql-connector
redis
